--- types/entities/Availability.js
@@ -1,8 +1,0 @@
-// @flow
-declare module 'brewskey.js-api' {
-  declare type Availability = {
-    description: ?string,
-    id: string,
-    name: string,
-  };
-}
