--- types/entities/Glass.js
@@ -1,10 +1,0 @@
-// @flow
-declare module 'brewskey.js-api' {
-  declare type Glass = {
-    createDate: Date,
-    description: ?string,
-    id: string,
-    name: string,
-    updateDate: Date,
-  };
-}
