--- types/entities/Pour.js
@@ -1,25 +1,0 @@
-// @flow
-declare module 'brewskey.js-api' {
-  declare type Pour = {
-    beverage: {
-      id: number,
-      name: string,
-    },
-    id: number,
-    location: {
-      id: number,
-      name: string,
-    },
-    ounces: number,
-    owner: {
-      id: string,
-      userName: string,
-    },
-    pourDate: string,
-    pulses: number,
-    tap: {
-      id: number,
-      name: string,
-    },
-  };
-}
