--- types/entities/Srm.js
@@ -1,8 +1,0 @@
-// @flow
-declare module 'brewskey.js-api' {
-  declare type Srm = {
-    hex: string,
-    id: string,
-    name: string,
-  };
-}
