--- types/entities/Style.js
@@ -1,20 +1,0 @@
-// @flow
-declare module 'brewskey.js-api' {
-  declare type Style = {
-    abvMax: number,
-    abvMin: number,
-    createDate: Date,
-    description: ?string,
-    fgMax: number,
-    fgMin: number,
-    ibuMax: number,
-    ibuMin: number,
-    id: string,
-    name: string,
-    ogMax: number,
-    ogMin: number,
-    srmMax: number,
-    srmMin: number,
-    updateDate: Date,
-  };
-}
